refactor: rename `:feature:search` to `:feature:searchresult`
In preparation for the search screen rewrite.
