+++ app/src/main/kotlin/com/potatosheep/kite/app/ui/KiteApp.kt
@@ -25,7 +25,7 @@ import com.potatosheep.kite.feature.bookmark.impl.navigation.bookmarkEntry
 import com.potatosheep.kite.feature.image.navigation.imageEntry
 import com.potatosheep.kite.feature.onboarding.impl.navigation.onboardingEntry
 import com.potatosheep.kite.feature.post.impl.navigation.postEntry
-import com.potatosheep.kite.feature.search.impl.navigation.searchEntry
+import com.potatosheep.kite.feature.searchresult.impl.navigation.searchEntry
 import com.potatosheep.kite.feature.settings.impl.navigation.settingsEntry
 import com.potatosheep.kite.feature.subreddit.impl.navigation.subredditEntry
 import com.potatosheep.kite.feature.user.impl.navigation.userEntry
