--- feature/search/api/build.gradle.kts
@@ -1,12 +1,0 @@
-plugins {
-    alias(libs.plugins.kite.android.feature.api)
-}
-
-android {
-    namespace = "com.potatosheep.kite.feature.search.api"
-}
-
-dependencies {
-    implementation(projects.core.common)
-    api(projects.core.navigation)
-}
