+++ feature/searchresult/impl/.gitignore
