+++ feature/searchresult/impl/src/main/AndroidManifest.xml
