+++ feature/searchresult/impl/src/main/res/mipmap-hdpi/ic_launcher_round.webp
