+++ feature/searchresult/impl/src/main/res/mipmap-xhdpi/ic_launcher_round.webp
