+++ feature/searchresult/impl/src/main/res/mipmap-xxxhdpi/ic_launcher.webp
