+++ feature/searchresult/impl/src/main/res/values/colors.xml
