+++ feature/searchresult/impl/src/main/res/values/strings.xml
