+++ feature/subreddit/impl/build.gradle.kts
@@ -23,7 +23,7 @@ dependencies {
     implementation(projects.feature.post.api)
     implementation(projects.feature.user.api)
     implementation(projects.feature.image.api)
-    implementation(projects.feature.search.api)
+    implementation(projects.feature.searchresult.api)
     implementation(projects.feature.video.api)
 
     implementation(libs.coil)
