+++ feature/user/impl/src/main/kotlin/com/potatosheep/kite/feature/user/impl/navigation/UserEntryProvider.kt
@@ -7,7 +7,7 @@ import com.potatosheep.kite.core.designsystem.defaultTransitionSpec
 import com.potatosheep.kite.core.navigation.Navigator
 import com.potatosheep.kite.feature.image.api.navigation.navigateToImage
 import com.potatosheep.kite.feature.post.api.navigation.navigateToPost
-import com.potatosheep.kite.feature.search.api.navigation.navigateToSearch
+import com.potatosheep.kite.feature.searchresult.api.navigation.navigateToSearch
 import com.potatosheep.kite.feature.subreddit.api.navigation.navigateToSubreddit
 import com.potatosheep.kite.feature.user.api.navigation.UserNavKey
 import com.potatosheep.kite.feature.user.impl.UserRoute
