+++ settings.gradle.kts
@@ -45,8 +45,8 @@ include(":core:common")
 include(":feature:subredditwiki")
 include(":feature:user:impl")
 include(":feature:user:api")
-include(":feature:search:impl")
-include(":feature:search:api")
+include(":feature:searchresult:impl")
+include(":feature:searchresult:api")
 include(":core:markdown")
 include(":feature:video:impl")
 include(":feature:video:api")
